pyproject.toml
[project]
name = "time-frame"
version = "0.1.0"
description = ""
authors = [
    {name = "XetPy1030",email = "xetpy1030@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "numpy (>=2.2.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
